split a webapp module for jetty-ee10-websocket-jakarta-client
