+++ jetty-ee10/jetty-ee10-websocket/jetty-ee10-websocket-jakarta-client-webapp/src/main/java/org/eclipse/jetty/ee10/websocket/jakarta/client/webapp/JakartaWebSocketShutdownContainer.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.ee10.websocket.jakarta.client;
+package org.eclipse.jetty.ee10.websocket.jakarta.client.webapp;
 
 import java.util.Set;
 
@@ -20,6 +20,7 @@ import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
 import jakarta.servlet.ServletException;
+import org.eclipse.jetty.ee10.websocket.jakarta.client.JakartaWebSocketClientContainer;
 import org.eclipse.jetty.util.component.ContainerLifeCycle;
 import org.eclipse.jetty.util.component.LifeCycle;
 import org.slf4j.Logger;
