+++ jetty-ee10/jetty-ee10-websocket/jetty-ee10-websocket-jakarta-client-webapp/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,0 +1,1 @@
+org.eclipse.jetty.ee10.websocket.jakarta.client.webapp.JakartaWebSocketShutdownContainer
