--- jetty-ee10/jetty-ee10-websocket/jetty-ee10-websocket-jakarta-client/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,0 @@
-org.eclipse.jetty.ee10.websocket.jakarta.client.JakartaWebSocketShutdownContainer
